+++ packages/ricos-styles/web/src/index.ts
@@ -1,2 +1,3 @@
 export { StylesContext } from './context';
 export { default as Styles } from './styles';
+export { default as DocumentStyle } from './document-style/document-style';